refactor(virrun): the shared types and the launcher name get their own file
Three types that other modules import were sharing a file with the shape
they were written for, and the Windows launcher every WSL command goes
through was a bare string in six of them.

Closes the six virrun units of the file-organization sweep.
