+++ packages/virrun/src/models/exec/ExecStdio.ts
@@ -1,0 +1,2 @@
+// "pipe" captures stdout/stderr into the ExecResult; "inherit" streams them live to the host terminal.
+export type ExecStdio = "inherit" | "pipe";
