+++ packages/virrun/src/models/exec/ExecTeeTarget.ts
@@ -1,0 +1,2 @@
+// The host stream a tee mirrors the child's stdout onto.
+export type ExecTeeTarget = "stderr" | "stdout";
