+++ packages/virrun/src/models/exec/FlushOpType.ts
@@ -1,0 +1,7 @@
+// One host-side operation in a write-back flush plan (buildFlushPlan, specs/write-back.md).
+export enum FlushOpType {
+  // Copy the upper entry onto the host (mkdir for a dir, overwrite for a file).
+  Copy = "copy",
+  // Remove the host path recursively — a deletion, or clearing an opaque dir.
+  Delete = "delete",
+}
