+++ packages/virrun/src/models/exec/cache/TaskCacheEntry.ts
@@ -1,6 +1,6 @@
 import type { FlushOp } from "#src/models/exec/FlushOp";
 
-import { FlushOpType } from "#src/models/exec/FlushOp";
+import { FlushOpType } from "#src/models/exec/FlushOpType";
 import { createUniqueArraySchema } from "@esposter/shared";
 import { z } from "zod";
 // The recorded outcome of one exit-0 persist run, replayed verbatim on a later cache hit so the sandbox is skipped
