+++ packages/virrun/src/services/exec/cache/parseTaskCacheEntry.test.ts
@@ -1,4 +1,4 @@
-import { FlushOpType } from "#src/models/exec/FlushOp";
+import { FlushOpType } from "#src/models/exec/FlushOpType";
 import { parseTaskCacheEntry } from "#src/services/exec/cache/parseTaskCacheEntry";
 import { TEST_FILENAME } from "#src/services/exec/util/constants.test";
 import { describe, expect, test } from "vitest";
