+++ packages/virrun/src/services/exec/snapshot/removeSnapshotDirectoriesDetached.ts
@@ -1,7 +1,7 @@
 import { writeVirrunDebug } from "#src/services/cli/debug/writeVirrunDebug";
 import { removeSnapshotDirectory } from "#src/services/exec/snapshot/removeSnapshotDirectory";
 import { spawnBackground } from "#src/services/exec/util/spawnBackground";
-import { VIRRUN_REMOVE_LIST_TEMP_PREFIX, WSL_REMOVE_LIST_SCRIPT, WSL_UNC_REGEX } from "#src/services/exec/wsl/constants";
+import { VIRRUN_REMOVE_LIST_TEMP_PREFIX, WSL_EXECUTABLE, WSL_REMOVE_LIST_SCRIPT, WSL_UNC_REGEX } from "#src/services/exec/wsl/constants";
 import { getWslNativeCacheRoot } from "#src/services/exec/wsl/getWslNativeCacheRoot";
 import { joinNullDelimited } from "#src/services/exec/wsl/joinNullDelimited";
 import { readWslPath } from "#src/services/exec/wsl/readWslPath";
@@ -53,7 +53,7 @@ export const removeSnapshotDirectoriesDetached = (dirs: readonly string[]): void
     const listFilename = `${VIRRUN_REMOVE_LIST_TEMP_PREFIX}${process.pid}.${crypto.randomUUID()}`;
     const listUnc = join(cacheRoot, listFilename);
     writeFileSync(listUnc, joinNullDelimited(linuxDirs));
-    spawnBackground("wsl.exe", ["--exec", "sh", "-c", WSL_REMOVE_LIST_SCRIPT, "sh", readWslPath(listUnc)]);
+    spawnBackground(WSL_EXECUTABLE, ["--exec", "sh", "-c", WSL_REMOVE_LIST_SCRIPT, "sh", readWslPath(listUnc)]);
   }).match(noop, ({ message }) => {
     writeVirrunDebug(`detached teardown of ${linuxDirs.length} dirs not staged — ${message}`);
   });
